# backend/docs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ArchiTrack API",
    "version": "1.0.0",
    "description": "ArchiTrack Backend REST API Documentation",
    "contact": {
      "name": "ArchiTrack Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.architrack.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      }
    },
...
  },
...
  "paths": {
...
    "/api/site-surveys/images/{imageId}/annotations": {
      "get": {
        "summary": "注釈データ取得",
        "description": "画像に関連付けられた注釈データを取得する",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "画像ID"
          }
        ],
        "responses": {
          "200": {
            "description": "注釈データ取得成功（注釈が存在しない場合はdata: nullを返す）",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "imageId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "data": {
                          "type": "object"
                        },
                        "version": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "無効な画像ID形式"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "画像が見つからない"
          }
        }
      },
      "put": {
        "summary": "注釈データ保存",
        "description": "画像に注釈データを保存する（新規作成または更新）",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "画像ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "data"
                ],
                "properties": {
                  "data": {
                    "type": "object",
                    "required": [
                      "objects"
                    ],
                    "properties": {
                      "version": {
                        "type": "string"
                      },
                      "objects": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      },
                      "background": {
                        "type": "string"
                      },
                      "viewportTransform": {
                        "type": "array",
                        "items": {
                          "type": "number"
                        }
                      }
                    }
                  },
                  "expectedUpdatedAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "楽観的排他制御用の期待される更新日時"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注釈データ保存成功"
          },
          "400": {
            "description": "バリデーションエラーまたは無効な注釈データ形式"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "画像が見つからない"
          },
          "409": {
            "description": "楽観的排他制御エラー（競合）"
          }
        }
      }
    },
    "/api/site-surveys/images/{imageId}/annotations/export": {
      "get": {
        "summary": "注釈データJSONエクスポート",
        "description": "注釈データをJSONファイルとしてエクスポートする",
        "tags": [
          "Annotations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "画像ID"
          }
        ],
        "responses": {
          "200": {
            "description": "JSONエクスポート成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "無効な画像ID形式"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "画像または注釈データが見つからない"
          }
        }
      }
    },
...
  }
}